9.2 trinkets/lower_opt_cd/hidden
One of them... doesn't have a cooldown? Might remove it later, let's see for now.
The more interesting one is the resonator. It pretends it has a "duration" of 4s.

The previous logic was poorly done, and I think was only working as a combination of charges+opt_lower_cooldown (which is... rare, to say the least).
The new version should work fine in general cases.

Add a new `hidden` property on spell data so gex can decide what spells to hide.
While you may think "Well, just hide spells without a cooldown!", this sadly doesn't work with i.e. Adaptation, or potentially with the new 9.2 PvP trinkets.

The other possible solution was a flag `.force_show`, I don't think it really matters.